requirements.txt
supabase
python-dotenv
openai
vecs